Added mixed precision, XLA and larger batch for faster training
Speed package (safe — does not change the trained weights' accuracy target,
only throughput):

- Mixed precision (float16): set_global_policy("mixed_float16") in __main__,
  controlled by SNC_MIXED_PRECISION (default on). 1.5-2x faster on Tensor
  Core GPUs (T4/A100/L4) and halves activation memory. Output layers
  (target_mask, class_presence) and the mask-apply Multiply are pinned to
  dtype="float32" so the sigmoid outputs and L1/BCE losses stay numerically
  stable. Adam is wrapped in LossScaleOptimizer under mixed precision
  (guarded — falls back to plain Adam on any failure or under float32).

- XLA: jit_compile passed to model.compile, controlled by SNC_XLA
  (default on). Fuses the train step; fails fast at step 1 if an op can't
  compile, so set SNC_XLA=0 to fall back.

- Larger batch: SNC_BATCH_SIZE (default 32, was 16) keeps a fast GPU busy.

All three are env-toggleable with no code edits. Training logs the active
config (mixed precision policy, jit_compile, batch size) and records it in
the Drive run-log metrics.

Notebook: updated GPU guidance — L4/A100 are now recommended over T4
(TF 2.20 ships their kernels, so CUDA_ERROR_INVALID_PTX no longer applies);
documented the three speed env vars.

+++ src/model_training/conditioned_separator.py
@@ -139,8 +139,11 @@ def build_conditioned_separator(
     d1 = Conv2DTranspose(base_filters, 2, strides=2, padding="same")(d2)
     d1 = _conv_block(Concatenate()([d1, e1]), base_filters)           # (256, 128)
 
+    # dtype="float32" pins the output to full precision even under a global
+    # mixed_float16 policy: the sigmoid mask and the downstream loss must stay
+    # in float32 for numerical stability (float16 saturates/over-rounds here).
     mask = Conv2D(1, 1, padding="same", activation="sigmoid",
-                  name="target_mask")(d1)
+                  name="target_mask", dtype="float32")(d1)
 
     if with_detection_head:
         # Detection head: bottleneck is already FiLM-conditioned on the class
@@ -148,8 +151,9 @@ def build_conditioned_separator(
         # Dense(1, sigmoid) then predicts P(queried class present | mixture).
         det = GlobalAveragePooling2D(name="det_pool")(bottleneck)
         det = Dense(128, activation="relu", name="det_dense")(det)
+        # float32 output for stable binary cross-entropy under mixed precision.
         class_presence = Dense(1, activation="sigmoid",
-                               name="class_presence")(det)
+                               name="class_presence", dtype="float32")(det)
         return Model([log_mag, query], [mask, class_presence],
                      name=f"conditioned_separator_f{base_filters}_det")
